Update : Observer 패턴 구현
