+++ src/components/Element.js
@@ -1,8 +1,21 @@
 export default class Element{
-  #template = null;
-  constructor(){
+  constructor(model){
+    console.log('model', model);
+    this.formStateModel = model;
   }
   getTemplate(){
     return this.template;
+  }
+  render(parent){
+    console.log(parent, this.template);
+    document.querySelector(parent).append(this.template);
+  }
+  convertElement(templateStr){
+    const div = document.createElement('div');
+    div.innerHTML = templateStr.trim();
+    return div.firstChild;
+  }
+  setOnClickListener($target, clickEvent){
+    $target.addEventListener('click', clickEvent);
   }
 }
