+++ src/stores/index.js
@@ -1,0 +1,11 @@
+export default class Observer {
+  constructor() {
+    this._observers = new Set();
+  }
+  subscribe(observer) {
+    this._observers.add(observer);
+  }
+  notify() {
+    this._observers.forEach((observer) => observer());
+  }
+}
